Make EFI boot work in all cases
Previously, the isohybrid image could boot on EFI firmware as well as
non-EFI firmware when written to an ISO image, and on non-EFI firmware
when dd'ed to USB mass storage, but not on a system with EFI firmware.

Add a call "isohybrid -u" at the end of the script to twiddle the magic
bits so this becomes possible again.

Thanks to Thomas Schmitt for guidance.