Use gpgv1.4-udeb for orion5x network-console
gpgv-udeb used to be built from the gnupg (1.4) package but is now built
from gnupg2.  This brought in a dependency on libgcrypt20 which is huge
and caused the QNAP TS-09x images to break due to size issues.  We
created a gpgv1.4-udeb on armel built from gnupg 1.4 to fix the orion5x
network-console image.  Other options will be explored (see #814027).