Provide kernel and ramdisk for kirkwood netboot
Make the kernel (without any DTB additions) and the installer ramdisk
available.  This allows people to easily append the DTB to the kernel
for devices which are unsupported or not fully supported.

While this practice is not very common on armel/kirkwood (unlike
on armhf), it's good to make it easy for expert users.  Thanks to
Adam Baker making this suggestion.