Generate image for Seagate Personal Cloud and Seagate NAS
This adds an uImage and uInitrd for Seagate devices, specifically
the Personal Cloud and Seagate NAS.  The uImage is just the kernel
as an u-boot image with the correct load address.  The DTB needs
to be loaded separately.  The uInitrd is an u-boot image of the
installer initrd to which an initrd is appended containing a preseed
file for noninteractive network configuration for network-console.