Fix Built-Using generation.
Let's build the “source (= version)” mapping correctly, which means
taking into account a possible version in the Source field. Otherwise,
take the Source field and the binary version. If there's no Source
field, take the binary name and version.

Otherwise, we might be getting a REJECT from ftp-master for unknown
packages in Built-Using.