Ensure we don't have libcrypto1.1 on orion5x-qnap
The orion5x-qnap image was broken again because of size issues.
This time, the problem was that both libcrypto1.0.2-udeb and
libcrypto1.1-udeb got pulled in.  Fortunately, the latter is
only used by wpasupplicant-udeb which we don't need on
orion5x-qnap.

Exclude wpasupplicant-udeb and libcrypto1.1-udeb.  If other udebs
will start moving to libcrypto1.1-udeb, this will break but
hopefully this won't happen for stretch and orion5x-qnap will be
removed after stretch anyway.