Add support for the unreleased suite (Closes: #852215).
This suite is used by non-released ports to carry packages or changes
which don't really belong to the Debian archive. Set USE_UNRELEASED=1
and USE_UDEBS_FROM=unstable when not building for a release arch.

This means we have to keep track of an hardcoded list of architectures
but hopefully we'll live with it.

Inspired-by: James Clarke <jrtc27@debian.org>
Signed-off-by: Cyril Brulebois <kibi@debian.org>