Increase the partition size for the armhf SD card images (Closes: #864260).
The installer binary size has grown over time and now exceeds the
predefined SD card partition size for the netboot build on armhf.
The hd-media image just tightly fits for now but will probably
outgrow the available space in the near future as well.

Increasing the partition size from ~40MB to ~50MB for both netboot
and hd-media would probably solve the issue for some time, but
during discussing the bug on debian-boot it has been proposed to
increase the size to ~100MB to ease the debugging process when
testing d-i images with alternate kernels.  As the resulting
increase in compressed binary size (~60kB) and write time to the
target (a few seconds) are insignificant and such a change makes
the debugging easier, the raw image size gets increased to ~100MB.