Add support for SNAPSHOT_TS
This makes it possible to use a specific timestamp of the Debian archive
on snapshot.debian.org instead of the detected mirror; disable
Valid-Until check when this is requested, since signatures are likely
expired. Most useful for bisecting when a specific regression was
introduced.

See: #872598

Note: This relies on apt (>= 1.1~exp9) for Check-Valid-Until support
inside sources.list. Otherwise, build/util/get-packages needs to be
modified to pass an extra option to apt-get.