mydistro: Add sources.list.udeb.local instead of generating from system
For old releases that no longer maintained on official mirrors we might
want to use archive.debian.org.

For new releases we might want to not to depend on build host system
configuration that could be inappropriate for debian-installer.

Note that only main/debian-installer for <release> and optionally from
<release>-proposed-updates should be specified in this file to avoid
build failures like following observed on <stretch> host:

  dpkg (subprocess): failed to chroot to './tmp/cdrom_isolinux/tree': Operation not permitted
  dpkg: error processing archive udebs/libc6.udeb (--unpack):
   subprocess new pre-installation script returned error exit status 2

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>