mydistro: Use wheezy specific installer vs installed options separator
In wheezy and earlier -- is used instead of --- as installed options
separator in boot entry. Installed options are copied from boot entry
to installed system boot entry.

While there remove nosmt kernel option as it is not supported by
wheezy linux kernel of 3.2 version. Main intend of this option is
to disable Hyper-Threading on Intel (R) CPUs that can be handy to
address L1TF and some workloads (e.g. software packet forwarding).

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>