Make efi-image generate a better-sized efi.img
The calculations for the required image size have always looked a
little suspect for me, only really designed for one file. Now we have
more files in the ESP, and potentially *many* more with DTBs too the
size calculated is much bigger than we need. Let's improve on that and
not waste space on our installation media, netboot downloads etc.

Refactor the code to do the following:

 * Do the current calculations as normal.
 * Make an image as we currently do.
 * Using fsck.msdos, find out exactly how much space we're wasting in
   the image we've just created,
 * Drop the chosen image size down appropriately and make the image
   again, leaving just a few kB spare.