Add wireless-regdb-udeb to Linux builds that include nic-wireless-modules
The Linux kernel uses wireless regulatory information to ensure that
wireless hardware only transmits on permitted frequencies and power
levels.  This information comes from some combination of AP beacons,
non-volatile memory on the wireless interface, a regulatory database
(regdb) file, and as a last resort a built-in "world regulatory
domain".

In the installer, we don't currently include regdb, and rely on the
other sources.  Older Linux kernel versions required extra user-space
software (CRDA) to load regdb, which was never packaged as a udeb
and probably wasn't worth the trouble.

However, the Linux kernel can now directly load and verify regdb, so
it's just a case of adding regdb and its detached signature to the
installer image.  Theefore add wireless-regdb-udeb to all the package
lists that include wifi drivers (nic-wireless-modules).