[arm64, armhf] Update SD-card-images sizes for netboot, netboot-gtk and hd-media.
The size for all images was previously 100MB. The arm64 netboot-gtk is
updated to 200MB and the other images 150MB.

The arm64 netboot-gtk images were nearly at capacity, in danger of
small changes breaking the image.

All images should have enough empty space to append all kernel modules
to the initrd when debugging and testing new platforms (currently 45MB
on armhf, and 56MB on arm64), and some room to grow above and
beyond. This change currently leaves free space of about 100MB on the
arm64 netboot-gtk image and 70MB on the other images.