Fix handling of speech installer menu entry in grub
The syslinux menu that appears in BIOS mode shows the "Install with
speech synthesis" menu entry only in the GTK variant of the mini.iso.

The grub menu showed that entry in all cases. Now we add a supplementary
test controlling the display of the menu entry, and it reuses the logic
shown in build/util/syslinux-cfgs:
* with IS_PURE_GTK set to 1, we have a menu entry pointing to
  ${INITRD} with the speech synthesis enabled because the default
  initrd is a GTK one
* when IS_PURE_GTK is not set, we add the menu entry only when we have
  a supplementary GTK initrd (${INITRD_GTK}) that can be used for the
  menu entry

Signed-off-by: Raphaël Hertzog <raphael@offensive-security.com>