Refactor depthcharge images into their own Makefile target
The netboot images for ChromeOS devices show up in the manifests as
"netboot/debian-installer/../depthcharge". This is because they were
implemented as part of the NETBOOT_DIR target which is about the
"debian-installer" directory (not the build flavour's "netboot" dir).

Refactor these images into a DEPTHCHARGE target, so we can add a cleaner
path to the manifests. This also seems more in line with how the build
system was designed to be extended.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>