Update translation-status for the release.
Cheat a little, patching calc-release-status to move stats/apt-setup.5 out
of the way before computing per-language stats.

The brand new apt-setup/non-free-firmware template means 2 extra msgids
(short and long descriptions), which appear fuzzy for almost everyone.

Flagging it with :sl5: made it easy to spot, to implement some targeting
fudging: that question only appears in expert mode, is almost identical to
the non-free one, and we expect people to read about the new component in
the release announcement, in the installation guide, in the release notes,
etc. There's no need to pretend some languages aren't fully translated
just because translators haven't had a chance to catch up with this new
template.