Make it possible to check version requirements.
Let's have some safety net when it comes to preparing releases: we want
to be able to upload shortly after dinstall has deployed the last set of
packages we worked on, but we don't have any guarantee all buildds will
see the updated archive within a given time window.

Since we're dealing with udebs fetched from testing, there's no way to
express that via Build-Depends. Let's have some dedicated check during
the build instead.

Since the list of package_version is only useful temporarily, and mainly
as a safety net, it is not planned to document CHECK_MINIMAL_VERSION
updates in the changelog.