Fix colors in GRUB submenus for x86 netboot (Closes: #1036771).
Behaviour seems different with at least arm64, be it using netboot or
netinst: some kind of reset seems to be happening. In other words, no
longer hardcoding colors in submenus would mean only keeping them on the
main menu, and in the dark menu.

Therefore, set a specific flag (IS_X86) and use it to determine whether
hardcoding should be skipped.

This means touching amd64 and i386, while leaving arm* (and ia64) alone.