allow for a cut-down mini.iso build (for salsa)
If one enables CI/CE for debian-installer, the resulting
artifacts (including the tarball) exceed salsa's size limit for
artifacts, and even if that were not the case, I'm not sure anyone would
be particularly interested in doing this very often, let alone on every
push.

On the other hand, it is quite useful to be able to generate a mini.iso,
so that one can then test it.

To that end, I've defined a `salsa-ci` build profile that trims the
build target down to just `build_netboot-gtk` (it may be interesting to
also do a text-mode one at some point), and makes it include the
resulting mini.iso in the job's artifacts.

In the `salsa-ci.yml` included in this MR, I have gitlab's CI run in one
of two modes: 1. a fairly conventional salsa-ci/pipeline, with tests
etc. (a commit to D-I itself cause this to run). 2. as something that's
been triggered in a parent-child pipeline. (a commit to a
udeb-gererating repo. can trigger this, after its `build` & `aptly` jobs
have succeeded)

When triggered in a parent-child (2.) pipeline, the salsa job prepares
both `$PRESEED` and `build/sources.list.udeb.local` in order to include
the version of whatever started the pipeline running to be built into
D-I (either at build time, or download time)

Having built the mini.iso, this `salsa-ci.yml` is able to trigger
test(s) on openQA.