set auto-install/cloak_initrd_preseed in mini-ISO
This asks the auto-install code to ignore /preseed.cfg for the purpose
of detecting preseeding.

That let us do tests of the code for handling the case when a user
selects an automatic install, but neglects to set a preseed url/file.

It will also allow others to customise our install media (by adding a
/preseed.cfg) without removing the prompt for missing preseeds.

(the code to look for this setting will of course need to be added to
preseed before this actually does any of that)